Room Marking
User can use "Room Marking" button to begin marking rooms. This works by clicking different points for each corner of a room, tracing it to form a polygon. If user is near the end point, and clicks, a polygon will form and the temporary lines will be removed. Logic for settings to label the room and change its color to come next.