Logic for deleting floors
Created a logic for removing floors. Only allows user to delete the uppermost and lowermost floor at a time, not including ground (main) floor. If the canvas attempting to be removed is empty, it will remove. If there are any objects, it will prompt the user with a popup confirming if they wish to delete. Whilst deleting the currently active floor, it will switch the user to the level below it whilst deleting.