Copying specific objects to new floors
Added logic for cloning specific objects to new floors. Use cases of this would include copying stairs and eventually wall bases. This will help with cohesiveness in structure (i.e. stairs would be in a constant location between floors). The base will be the outer walls, user defined. The idea is to have the user lock/unlock the base so it will copy to new floors and be unmovable until unlocking.