Locking Objects
Added the ability to lock objects. Locking objects prevents movement and transformation. It also clones the object and moves them to a new floor when created. This allows for: Preventing moving walls/bases, and keep structure and organization for outside walls. If the next floor has a differently sized wall, the user can simply unlock and change to the correct shape or size (or remove).