Generating images and zip files
Wrote code that generates a PNG for each floor as an image (hiding devices), storing them in a zip file. If user is choosing to export for home assistant, it will await JSON file generation and canvas image generation before creating the downloaded zip file. If a user is only exporting a singular file, it should save as a normal png.