Solidifying Structure and Code Refactoring
Stored important variable information in an object, labeled as the canvas data, canvas info, and canvas states. This helps simplify the important variables necessary to pass to children components (and there is less of a mess). Edited the floor array to update in real time, and structure session storage to store important variables on refresh rather than update (so real time updates do not have to continuously access session storage).  Refined structure of exportation process.