Fixed entity names and added image URLs
Changed entity to the correct corresponding value for dashboard export. Original was the ID value, but it should be the value that's set as the object's 'original name'. Since Home Assistant can allow storing images in a local folder, utilized this with the image generation to synchronize file names for image url.