Unit Testing - Components
Separated component logic into it's own file, exporting creating function and filter function. As the filter function returns the called create function, it is no longer necessary to use the create function in the main canvas logic. Instead, when a component button is pressed, it sends the canvas action to the filter function, which filters, creates the object, adds it to the canvas, and returns the result. A singular create instance is tested, and each canvasAction instance is also tested with the filter. Lastly, instances where canvas action is not component-based is tested (such as drawing a square).