Undo Feature
Added an undo feature with undo and redo buttons that appear with the object menu. CTRL Z and CTRL Y support is also added for a natural flow.
Canvas states and device states are stored in parallel based on each action. Undo or Redo changes the index for the array to change the current state. Changing in the middle of the state array will erase later states and begin from that point. A max limit of 20 states at a time is put in place, so older states will leave as new states come in to save on memory.